feat: refactor main.go into modular components with UI, database, and GitHub handlers
